Detect a variable if the expression is a leaf
The previous assumption that the evaluation_order include a bare
word to the list if an expression is a leaf is not true. In the
method _find_variables, detect whether an element is a leaf to
get the real variable.

This will correct an error wherein the interpreter throws an
error the bare word is passed to the boolean functions.